+++ src/components/addTable/AddTable.tsx
@@ -1,127 +1,145 @@
-import React, { useEffect, useRef } from 'react';
+import { useEffect, useState } from 'react';
+import { useScroll } from '../../hooks/useScroll';
 
-import { getTimeArray } from '../../utils/getTimeArray';
 import {
-  Blank,
-  Bottom,
-  Date,
-  DateWrapper,
-  Select,
-  SelectWrapper,
-  Time,
-  TimeWrapper,
-  Top,
-  Wrapper,
+  ButtonWrapper,
+  MoveButton,
+  ScrollbarThumb,
+  ScrollbarTrack,
+  TableWrapper,
 } from './AddTable.styles';
-import { AddTableType } from './AddTable.types';
-import Selecto from 'react-selecto';
-import _ from 'lodash';
+
+import addPrevDisable from '../../assets/icons/addPrevDisable.png';
+import addNextDisable from '../../assets/icons/addNextDisable.png';
+import addPrevActive from '../../assets/icons/addPrevActive.png';
+import addNextActive from '../../assets/icons/addNextActive.png';
+import Table from './Table';
+import { TableSelectedTypes } from './AddTable.types';
+import { getAddTimeTableInfo } from '../../utils/getAddTimeTableInfo';
+
+interface AddTableTypes {
+  selectedMethod: string;
+  dates: string[];
+  timeRange: number[];
+  previousSelectedTimes: string[];
+  tableSelected: TableSelectedTypes;
+  setTableSelected: React.Dispatch<React.SetStateAction<TableSelectedTypes>>;
+}
 
 const AddTable = ({
-  contentRef,
+  selectedMethod,
+  timeRange,
+  dates,
+  previousSelectedTimes,
   tableSelected,
   setTableSelected,
-  selectedMethod,
-  tablePage,
-  validDateChunks,
-  times,
-}: AddTableType) => {
-  const timeDetail = getTimeArray(times);
-  const selectoRef = useRef<any>(null);
+}: AddTableTypes) => {
+  const [tablePage, setTablePage] = useState(0);
+
+  const validDateChunks = getAddTimeTableInfo(dates);
 
   useEffect(() => {
-    if (tableSelected[tablePage]) {
-      tableSelected[tablePage].forEach((id) => {
-        const element = document.getElementById(id);
-        element?.classList.add('selected');
+    const newObj: TableSelectedTypes = {};
+
+    previousSelectedTimes.forEach((time: any) => {
+      validDateChunks.map((chunk, index) => {
+        chunk.map((date) => {
+          if (date.date === time.slice(0, 10)) {
+            if (newObj[index] === undefined) {
+              newObj[index] = [time];
+            } else {
+              newObj[index].push(time);
+            }
+          }
+        });
       });
-
-      selectoRef.current.setSelectedTargets(
-        tableSelected[tablePage].map((id) => document.getElementById(id))
-      );
-    }
-  }, [tablePage, selectedMethod, tableSelected]);
-
-  const handleCellSelect = (e: any) => {
-    e.added.forEach((el: any) => {
-      el.classList.add('selected');
     });
-
-    e.removed.forEach((el: any) => {
-      el.classList.remove('selected');
-    });
-  };
-
-  const addSelectedToObject = () => {
-    const newArr: string[] = Array.from(
-      document.querySelectorAll('.selected')
-    ).map((node: Element) => node.id);
-
-    const newObj = { ...tableSelected };
-    newObj[tablePage] = newArr;
 
     setTableSelected(newObj);
+  }, [previousSelectedTimes]);
+
+  useEffect(() => {
+    const newObj: TableSelectedTypes = {};
+
+    previousSelectedTimes.forEach((time) => {
+      validDateChunks.map((chunk, index) => {
+        chunk.map((date) => {
+          if (date.date === time.slice(0, 10)) {
+            if (newObj[index] === undefined) {
+              newObj[index] = [time];
+            } else {
+              newObj[index].push(time);
+            }
+          }
+        });
+      });
+    });
+
+    setTableSelected(newObj);
+  }, [previousSelectedTimes]);
+
+  const {
+    contentWrapperRef,
+    contentRef,
+    trackRef,
+    thumbRef,
+    offsetY,
+    handleMouseDown,
+    handleTouchStart,
+    handleDragStart,
+  } = useScroll();
+
+  const handlePrevButtonClick = () => {
+    if (tablePage !== 0) {
+      setTablePage(tablePage - 1);
+    }
+  };
+
+  const handleNextButtonClick = () => {
+    if (tablePage !== validDateChunks.length - 1) {
+      setTablePage(tablePage + 1);
+    }
   };
 
   return (
-    <Wrapper ref={contentRef}>
-      <Top>
-        <Blank />
-        <DateWrapper>
-          {validDateChunks[tablePage].map(
-            ({ date, isValidDate }: { date: string; isValidDate: boolean }) =>
-              date.slice(0, 5) === 'blank' ? (
-                <Date key={date} isValidDate={isValidDate} />
-              ) : (
-                <Date key={date} isValidDate={isValidDate}>{`${date.slice(
-                  5,
-                  7
-                )}월${date.slice(8, 10)}일`}</Date>
-              )
-          )}
-        </DateWrapper>
-      </Top>
-
-      <Bottom>
-        <TimeWrapper>
-          {times.map((time) => (
-            <Time key={time}>{time}</Time>
-          ))}
-        </TimeWrapper>
-
-        {validDateChunks[tablePage]?.map(
-          ({ date, isValidDate }: { date: string; isValidDate: boolean }) => (
-            <SelectWrapper key={date} className="container">
-              <Selecto
-                ref={selectoRef}
-                dragContainer={'.container'}
-                selectableTargets={['.valid']}
-                onSelect={handleCellSelect}
-                onDragEnd={addSelectedToObject}
-                hitRate={0}
-                selectByClick={true}
-                selectFromInside={true}
-                continueSelect={true}
-                continueSelectWithoutDeselect={false}
-                toggleContinueSelect={'shift'}
-                toggleContinueSelectWithoutDeselect={[['ctrl'], ['meta']]}
-                ratio={0}
-              ></Selecto>
-
-              {timeDetail.map((time) => (
-                <Select
-                  className={isValidDate ? 'valid' : 'invalid'}
-                  key={`${date} ${time}:00`}
-                  id={`${date} ${time}`}
-                  selectedMethod={selectedMethod}
-                  isValidDate={isValidDate}
-                />
-              ))}
-            </SelectWrapper>
-          )
-        )}
-      </Bottom>
-    </Wrapper>
+    <>
+      <ButtonWrapper>
+        <MoveButton
+          src={tablePage === 0 ? addPrevDisable : addPrevActive}
+          alt="Prev Button"
+          onClick={handlePrevButtonClick}
+        />
+        <MoveButton
+          src={
+            tablePage !== validDateChunks.length - 1
+              ? addNextActive
+              : addNextDisable
+          }
+          alt="Next Button"
+          onClick={handleNextButtonClick}
+        />
+      </ButtonWrapper>
+      <TableWrapper ref={contentWrapperRef}>
+        <Table
+          contentRef={contentRef}
+          tableSelected={tableSelected}
+          setTableSelected={setTableSelected}
+          times={timeRange}
+          tablePage={tablePage}
+          selectedMethod={selectedMethod}
+          validDateChunks={validDateChunks}
+        />
+      </TableWrapper>
+      <ScrollbarTrack ref={trackRef}>
+        <ScrollbarThumb
+          ref={thumbRef}
+          offsetY={offsetY}
+          onMouseDown={handleMouseDown}
+          onDragStart={handleDragStart}
+          onTouchStart={handleTouchStart}
+        />
+      </ScrollbarTrack>
+    </>
   );
 };
