+++ src/pages/addTime/AddTime.tsx
@@ -1,46 +1,32 @@
 import { useEffect, useRef, useState } from 'react';
+import { useNavigate, useParams } from 'react-router-dom';
 import { useRecoilState } from 'recoil';
 import _ from 'lodash';
 
 import { selectedMethodState } from '../../atoms/selectedMethodAtom';
-
-import addPrevDisable from '../../assets/icons/addPrevDisable.png';
-import addNextDisable from '../../assets/icons/addNextDisable.png';
-import addPrevActive from '../../assets/icons/addPrevActive.png';
-import addNextActive from '../../assets/icons/addNextActive.png';
+import { isTooltipShownState } from '../../atoms/isTooltipShownAtoms';
 
 import {
   Body,
-  ButtonWrapper,
   Main,
-  MoveButton,
-  ScrollbarThumb,
-  ScrollbarTrack,
-  TableWrapper,
   Title,
   TitleWrapper,
   Wrapper,
   CalendarWrapper,
 } from './AddTime.styles';
 import Header from '../../components/header/Header';
-import BottomButton from '../../components/bottomButton/BottomButton';
+import AddToggle from '../../components/addToggle/AddToggle';
 import AddTable from '../../components/addTable/AddTable';
 import AddCalendar from '../../components/addCalendar/AddCalendar';
-
-import AddToggle from '../../components/addToggle/AddToggle';
-import { RoomTypes } from '../../types/roomInfo';
-import { API } from '../../utils/API';
-import { useNavigate, useParams } from 'react-router-dom';
-import { getRange } from '../../utils/getRange';
-import { getAddTimeTableInfo } from '../../utils/getAddTimeTableInfo';
-import { useScroll } from '../../hooks/useScroll';
-import { getAllTimeRange } from '../../utils/getAllTimeRange';
+import BottomButton from '../../components/bottomButton/BottomButton';
 import Tooltip from '../../components/tooltip/Tooltip';
-import { isTooltipShownState } from '../../atoms/isTooltipShownAtoms';
 
-interface TableSelectedTypes {
-  [key: number]: string[];
-}
+import { RoomTypes } from '../../types/roomInfo';
+
+import { API } from '../../utils/API';
+import { getRange } from '../../utils/getRange';
+import { getAllTimeRange } from '../../utils/getAllTimeRange';
+import { TableSelectedTypes } from './AddTime.types';
 
 const AddTime = () => {
   const { roomUUID } = useParams();
@@ -48,7 +34,6 @@ const AddTime = () => {
 
   const wrapperRef = useRef<HTMLDivElement>(null);
 
-  const [currentRoomState, setCurrentRoomState] = useState<any>([]);
   const [room, setRoom] = useState<RoomTypes>({
     title: '',
     deadLine: null,
@@ -60,18 +45,18 @@ const AddTime = () => {
   });
 
   const { title, dates, startTime, endTime } = room;
+  const [timeRange, setTimeRange] = useState<number[]>([]);
 
   const isTableView = startTime !== null && endTime !== null ? true : false;
-  const [tablePage, setTablePage] = useState(0);
 
   const [selectedMethod, setSelectedMethod] =
     useRecoilState(selectedMethodState);
-  const [selected, setSelected] = useState<string[]>([]);
-  const [tableSelected, setTableSelected] = useState<TableSelectedTypes>({});
 
   const [previousSelectedTimes, setPreviousSelectedTimes] = useState<string[]>(
     []
   );
+  const [tableSelected, setTableSelected] = useState<TableSelectedTypes>({});
+  const [calendarSelected, setCalendarSelected] = useState<string[]>([]);
 
   const userName = localStorage.getItem('userName');
 
@@ -84,13 +69,6 @@ const AddTime = () => {
       setRoom(data);
     };
 
-    const getCurrentRoomInfo = async () => {
-      const { data } = await API.get(
-        `/api/room/${roomUUID}/available-time/group`
-      );
-      setCurrentRoomState(data);
-    };
-
     const getPreviousSelectedTimes = async () => {
       const { data } = await API.get(
         `/api/room/${roomUUID}/available-time?name=${userName}`
@@ -99,37 +77,12 @@ const AddTime = () => {
     };
 
     getRoomInfo();
-    getCurrentRoomInfo();
     getPreviousSelectedTimes();
   }, []);
 
-  const validDateChunks = getAddTimeTableInfo(dates);
-
-  useEffect(() => {
-    const newObj: TableSelectedTypes = {};
-
-    previousSelectedTimes.forEach((time) => {
-      validDateChunks.map((chunk, index) => {
-        chunk.map((date) => {
-          if (date.date === time.slice(0, 10)) {
-            if (newObj[index] === undefined) {
-              newObj[index] = [time];
-            } else {
-              newObj[index].push(time);
-            }
-          }
-        });
-      });
-    });
-
-    setTableSelected(newObj);
-  }, [previousSelectedTimes]);
-
-  const [times, setTimes] = useState<number[]>([]);
-
   useEffect(() => {
     if (startTime && endTime && wrapperRef.current) {
-      setTimes(
+      setTimeRange(
         getRange(parseInt(startTime.slice(0, 2)), parseInt(endTime.slice(0, 2)))
       );
 
@@ -138,26 +91,14 @@ const AddTime = () => {
     }
   }, [startTime, endTime]);
 
-  const allTimeRange = getAllTimeRange(dates, times);
-
-  const handlePrevButtonClick = () => {
-    if (tablePage !== 0) {
-      setTablePage(tablePage - 1);
-    }
-  };
-
-  const handleNextButtonClick = () => {
-    if (tablePage !== validDateChunks.length - 1) {
-      setTablePage(tablePage + 1);
-    }
-  };
-
   const goToCurrent = () => {
     document.body.style.overflow = '';
     (wrapperRef.current as HTMLDivElement).style.overflow = 'auto';
 
     navigate(`/current/${roomUUID}`);
   };
+
+  const allTimeRange = getAllTimeRange(dates, timeRange);
 
   const getPayload = async () => {
     if (selectedMethod === 'possible') {
@@ -170,7 +111,7 @@ const AddTime = () => {
         : {
             name: userName,
             hasTime: false,
-            availableDateTimes: [...selected],
+            availableDateTimes: [...calendarSelected],
           };
 
       const putAvailableTime = async () => {
@@ -196,8 +137,6 @@ const AddTime = () => {
           availableDateTimes: filteredTime,
         };
 
-        console.log('젭알', filteredTime);
-
         const putAvailableTime = async () => {
           await API.put(
             `/api/room/${roomUUID}/available-time`,
@@ -208,7 +147,8 @@ const AddTime = () => {
         putAvailableTime();
       } else {
         const newDates = dates.map((date) => `${date} 00:00`);
-        const filteredTime = selected && _.difference(newDates, selected);
+        const filteredTime =
+          calendarSelected && _.difference(newDates, calendarSelected);
 
         const payload = {
           name: userName,
@@ -232,19 +172,8 @@ const AddTime = () => {
     getPayload();
     goToCurrent();
 
-    // window.location.reload();
+    window.location.reload();
   };
-
-  const {
-    contentWrapperRef,
-    contentRef,
-    trackRef,
-    thumbRef,
-    offsetY,
-    handleMouseDown,
-    handleTouchStart,
-    handleDragStart,
-  } = useScroll();
 
   return (
     <Wrapper ref={wrapperRef}>
@@ -257,57 +186,27 @@ const AddTime = () => {
           <AddToggle
             isTableView={isTableView}
             setTableSelected={setTableSelected}
-            setSelected={setSelected}
+            setSelected={setCalendarSelected}
           />
           <Title>시간으로 선택해 주세요</Title>
         </TitleWrapper>
 
         <Main>
           {isTableView ? (
-            <>
-              <ButtonWrapper>
-                <MoveButton
-                  src={tablePage === 0 ? addPrevDisable : addPrevActive}
-                  alt="Prev Button"
-                  onClick={handlePrevButtonClick}
-                />
-                <MoveButton
-                  src={
-                    tablePage !== validDateChunks.length - 1
-                      ? addNextActive
-                      : addNextDisable
-                  }
-                  alt="Next Button"
-                  onClick={handleNextButtonClick}
-                />
-              </ButtonWrapper>
-              <TableWrapper ref={contentWrapperRef}>
-                <AddTable
-                  contentRef={contentRef}
-                  tableSelected={tableSelected}
-                  setTableSelected={setTableSelected}
-                  times={times}
-                  tablePage={tablePage}
-                  selectedMethod={selectedMethod}
-                  validDateChunks={validDateChunks}
-                />
-              </TableWrapper>
-              <ScrollbarTrack ref={trackRef}>
-                <ScrollbarThumb
-                  ref={thumbRef}
-                  offsetY={offsetY}
-                  onMouseDown={handleMouseDown}
-                  onDragStart={handleDragStart}
-                  onTouchStart={handleTouchStart}
-                />
-              </ScrollbarTrack>
-            </>
+            <AddTable
+              selectedMethod={selectedMethod}
+              dates={dates}
+              timeRange={timeRange}
+              previousSelectedTimes={previousSelectedTimes}
+              tableSelected={tableSelected}
+              setTableSelected={setTableSelected}
+            />
           ) : (
             <CalendarWrapper>
               <AddCalendar
                 dates={dates}
-                selected={selected}
-                setSelected={setSelected}
+                selected={calendarSelected}
+                setSelected={setCalendarSelected}
                 selectedMethod={selectedMethod}
               />
             </CalendarWrapper>
